Refactor the overall code according to code reviews
