+++ src/component/Buttons.jsx
@@ -1,13 +1,16 @@
 import React from 'react';
+import { increaseBySteps } from '../utills/increaseFunction';
 
 import Button from './Button';
 
-export default function Buttons({ buttons, onClick }) {
+export default function Buttons({ numbers, onClick }) {
   return (
     <div>
       {
-        buttons.map(({ index, increaseByStep }) => (
-          <Button value={index} onClick={() => onClick(increaseByStep)} />
+        numbers.map((i) => (
+          <Button onClick={() => onClick(increaseBySteps(i))}>
+            {i}
+          </Button>
         ))
       }
     </div>
